+++ confiture-rest-api/src/audits/audit.service.ts
@@ -1140,6 +1140,7 @@ export class AuditService {
         auditType: true,
         editUniqueId: true,
         consultUniqueId: true,
+        initiator: true,
         pages: {
           select: {
             results: true
@@ -1196,6 +1197,8 @@ export class AuditService {
         );
       }
 
+      const statementIsPublished = !!a.initiator;
+
       return {
         ...pick(
           a,
@@ -1207,7 +1210,8 @@ export class AuditService {
         ),
         complianceLevel,
         status: progress < 1 ? "IN_PROGRESS" : "COMPLETED",
-        estimatedCsvSize: 502 + a.pages.length * 318
+        estimatedCsvSize: 502 + a.pages.length * 318,
+        statementIsPublished
       };
     });
   }