+++ confiture-web-app/src/router.ts
@@ -25,6 +25,7 @@ import PrivacyPage from "./pages/misc/PrivacyPage.vue";
 import SiteMapPage from "./pages/misc/SiteMapPage.vue";
 import ContextPage from "./pages/report/ContextPage.vue";
 import ReportPage from "./pages/report/ReportPage.vue";
+import StatementPage from "./pages/StatementPage.vue";
 import RoadmapPage from "./pages/RoadmapPage.vue";
 import { useAccountStore, useAuditStore } from "./store";
 
@@ -245,7 +246,7 @@ const router = createRouter({
     },
     // Report pages
     {
-      path: "/rapports/:uniqueId/contexte",
+      path: "/rapport/:uniqueId/contexte",
       name: "context",
       component: ContextPage,
       meta: {
@@ -253,13 +254,46 @@ const router = createRouter({
         hideHomeLink: true
       }
     },
+    // TODO: remove this redirect in few months (17/04/2024)
     {
-      path: "/rapports/:uniqueId/:tab?",
+      path: "/rapports/:uniqueId/contexte",
+      name: "context-old",
+      redirect: () => {
+        return { name: "context" };
+      }
+    },
+    {
+      path: "/rapport/:uniqueId/:tab?",
       name: "report",
       component: ReportPage,
       meta: {
         name: "Rapport d’audit",
         hideHomeLink: true
+      }
+    },
+    // TODO: remove this redirect in few months (17/04/2024)
+    {
+      path: "/rapports/:uniqueId/:tab?",
+      name: "report-old",
+      redirect: () => {
+        return { name: "report" };
+      }
+    },
+    // a11y statement
+    {
+      path: "/declaration/:uniqueId",
+      name: "a11y-statement",
+      component: StatementPage,
+      meta: {
+        name: "Déclaration d’accessibilité"
+      }
+    },
+    // TODO: remove this redirect in few months (17/04/2024)
+    {
+      path: "/declarations/:uniqueId",
+      name: "a11y-statement-old",
+      redirect: () => {
+        return { name: "a11y-statement" };
       }
     },
     // Roadmap